backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
flask-jwt-extended
bcrypt
flask-restx
flask-cors
